purge old transformer code
