+++ impact/model/distgen_config.py
@@ -433,44 +433,3 @@ def make_variables(
             )
 
     return DistgenVariableMappings(input_mappings=mappings)
-
-
-def make_transformer(
-    var_mappings: DistgenVariableMappings,
-    config: DistgenVariableMappingConfig = DistgenVariableMappingConfig(),
-    transformer=None,
-):
-    """Build a transformer for a distgen Generator.
-
-    Parameters
-    ----------
-    var_mappings :
-        Pre-computed variable mappings from ``make_variables``.
-    config :
-        Variable mapping configuration.
-    transformer :
-        Optional existing transformer to extend.
-
-    Returns
-    -------
-    RoutingTransformer
-    """
-    from impact.model.transformer.routing import RoutingTransformer
-
-    if transformer is None:
-        _trans = RoutingTransformer()
-    else:
-        _trans = transformer
-
-    key_map = {m.var.name: m.key for m in var_mappings.input_mappings}
-
-    _trans.register_getter(
-        lambda gen, varname, _m=key_map: gen[_m[varname]],
-        regex=r"(?P<varname>.+)",
-    )
-    _trans.register_setter(
-        lambda gen, value, varname, _m=key_map: gen.__setitem__(_m[varname], value),
-        regex=r"(?P<varname>.+)",
-    )
-
-    return _trans
