--- impact/model/transformer/routing.py
@@ -1,169 +1,0 @@
-import re
-from typing import Any, Callable
-
-from impact.model.transformer.base import Transformer
-
-
-_Route = tuple[str, re.Pattern, list[str], Callable]
-
-
-def _pattern_to_regex(pattern: str) -> tuple[re.Pattern, list[str]]:
-    """Convert a pattern like 'quad_{name}_k1' to a compiled regex and list of param names."""
-    parts = re.split(r"\{(\w+)\}", pattern)
-    regex_parts = []
-    param_names = []
-    for i, part in enumerate(parts):
-        if i % 2 == 0:
-            regex_parts.append(re.escape(part))
-        else:
-            param_names.append(part)
-            regex_parts.append(f"(?P<{part}>.+?)")
-    return re.compile("^" + "".join(regex_parts) + "$"), param_names
-
-
-def _specificity(pattern: str, param_names: list[str]) -> tuple:
-    """Fewer variables = more specific; longer pattern breaks ties."""
-    return (len(param_names), -len(pattern))
-
-
-class RoutingTransformer(Transformer):
-    """
-    Routes get/set calls to registered handlers by pattern matching, similar to Flask/FastAPI.
-
-    Usage
-    -----
-    transformer = RoutingTransformer()
-
-    @transformer.setter(pattern="quad_{name}_k1")
-    def set_quad_k1(tool, value, name):
-        tool.ele[name]["k1"] = value
-
-    @transformer.getter(pattern="quad_{name}_k1")
-    def get_quad_k1(tool, name):
-        return tool.ele[name]["k1"]
-
-    transformer.set_property(sim, "quad_Q1_k1", 0.5)
-    transformer.get_property(sim, "quad_Q1_k1")
-
-    Pattern variables (e.g. `{name}`) are extracted from the property name and passed
-    as keyword arguments to the handler. More specific patterns (fewer variables, longer
-    literal portions) take priority over less specific ones.
-    """
-
-    def __init__(self):
-        self._setters: list[_Route] = []
-        self._getters: list[_Route] = []
-
-    # ------------------------------------------------------------------
-    # General register / decorator API
-
-    def register_setter(
-        self,
-        func: Callable,
-        *,
-        pattern: str | None = None,
-        regex: str | re.Pattern | None = None,
-    ) -> Callable:
-        """Register a setter function. Exactly one of *pattern* or *regex* must be given.
-
-        Parameters
-        ----------
-        pattern :
-            Template string like ``'quad_{name}_k1'``; ``{var}`` tokens become
-            named regex groups and are passed as kwargs to *func*.
-        regex :
-            A raw regex string or compiled ``re.Pattern``. Named groups are
-            extracted and passed as kwargs to *func*.
-        """
-        if (pattern is None) == (regex is None):
-            raise ValueError("Exactly one of 'pattern' or 'regex' must be provided")
-        if pattern is not None:
-            compiled, params = _pattern_to_regex(pattern)
-            sort_key = pattern
-        else:
-            compiled = re.compile(regex) if isinstance(regex, str) else regex
-            params = list(compiled.groupindex.keys())
-            sort_key = compiled.pattern
-        self._setters.append((sort_key, compiled, params, func))
-        self._setters.sort(key=lambda r: _specificity(r[0], r[2]))
-        return func
-
-    def register_getter(
-        self,
-        func: Callable,
-        *,
-        pattern: str | None = None,
-        regex: str | re.Pattern | None = None,
-    ) -> Callable:
-        """Register a getter function. Exactly one of *pattern* or *regex* must be given.
-
-        Parameters
-        ----------
-        pattern :
-            Template string like ``'quad_{name}_k1'``; ``{var}`` tokens become
-            named regex groups and are passed as kwargs to *func*.
-        regex :
-            A raw regex string or compiled ``re.Pattern``. Named groups are
-            extracted and passed as kwargs to *func*.
-        """
-        if (pattern is None) == (regex is None):
-            raise ValueError("Exactly one of 'pattern' or 'regex' must be provided")
-        if pattern is not None:
-            compiled, params = _pattern_to_regex(pattern)
-            sort_key = pattern
-        else:
-            compiled = re.compile(regex) if isinstance(regex, str) else regex
-            params = list(compiled.groupindex.keys())
-            sort_key = compiled.pattern
-        self._getters.append((sort_key, compiled, params, func))
-        self._getters.sort(key=lambda r: _specificity(r[0], r[2]))
-        return func
-
-    def setter(
-        self,
-        *,
-        pattern: str | None = None,
-        regex: str | re.Pattern | None = None,
-    ) -> Callable:
-        """Decorator sugar for register_setter."""
-
-        def decorator(func: Callable) -> Callable:
-            return self.register_setter(func, pattern=pattern, regex=regex)
-
-        return decorator
-
-    def getter(
-        self,
-        *,
-        pattern: str | None = None,
-        regex: str | re.Pattern | None = None,
-    ) -> Callable:
-        """Decorator sugar for register_getter."""
-
-        def decorator(func: Callable) -> Callable:
-            return self.register_getter(func, pattern=pattern, regex=regex)
-
-        return decorator
-
-    def _match(
-        self, registry: list[_Route], name: str
-    ) -> tuple[Callable, dict] | tuple[None, None]:
-        for _, regex, _, func in registry:
-            m = regex.match(name)
-            if m:
-                return func, m.groupdict()
-        return None, None
-
-    def set_property(self, tool: Any, name: str, value: Any) -> None:
-        """Route to the most specific registered setter matching `name`."""
-        func, kwargs = self._match(self._setters, name)
-        if func is None:
-            raise KeyError(f"No setter registered for '{name}'")
-        func(tool, value, **kwargs)
-
-    def get_property(self, tool: Any, name: str) -> Any:
-        """Route to the most specific registered getter matching `name`."""
-        func, kwargs = self._match(self._getters, name)
-        if func is None:
-            raise KeyError(f"No getter registered for '{name}'")
-        return func(tool, **kwargs)
